requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=1.4.23,<2.0.0
psycopg2-binary>=2.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
email-validator>=1.1.3
pydantic>=1.8.2,<2.0.0
python-dateutil>=2.8.2
alembic>=1.7.3
pytest>=6.2.5
httpx>=0.19.0
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
pytest-env>=0.6.2
pytest-mock>=3.6.1
pytest-xdist>=2.4.0
pylint>=2.11.1
black>=21.9b0
isort>=5.9.3
mypy>=0.910,<1.0
pre-commit>=2.15.0
sqlmodel>=0.0.8
aiofiles>=22.1.0